+++ changelog.md
@@ -1,10 +1,16 @@
 [0.0.1] 2022-02-07
 by esther615kim
+### ts-jest setup
 - `installed`
 - npx create-react-app my-app --template typescript
 - npm install --save-dev jest ts-jest @types/jest
 - npx ts-jest config:init
+- npm install --save-dev @testing-library/react @testing-library/user-event @testing-library/dom @testing-library/jest-dom
+- npx eslint --init
 
 - `added`
 -
 -
+
+- `changed`
+- .eslintrc.js add  "plugin:@typescript-eslint/recommended-requiring-type-checking"